Rename the `lexemes` module to `defaults`, since that's what it's providing.
As the Lex API in lrpar evolves, lrlex might choose to provide other
defaults beyond just lexemes, so tweaking the module name feels
sensible.
